+++ api/src/main/scala/pl/touk/nussknacker/engine/api/typed/typing.scala
@@ -131,18 +131,26 @@ object typing {
       } else if (klass.isPrimitive) {
         TypedClass(ClassUtils.primitiveToWrapper(klass), parameters)
       } else if (klass.isArray) {
-        //to not have separate class for each array, we pass Array of Objects
-        if (klass.getComponentType == classOf[Object]) {
-          TypedClass(klass, parameters)
-        } else parameters match {
-          case Nil =>
-            Typed.typedClass(classOf[Array[Object]], List(Typed(klass.getComponentType)))
-          case _: List[TypingResult] =>
-            throw new IllegalArgumentException(s"Array parameter passed twice, klass component type: ${klass.getComponentType}, type passed from parameters: ${parameters.head.display}")
-        }
+        decodeArrayType(klass, parameters)
       } else {
         TypedClass(klass, parameters)
       }
+
+    private def decodeArrayType(klass: Class[_], parameters: List[TypingResult]): TypedClass = {
+      val decodedComponentType = Typed(klass.getComponentType)
+      //to not have separate class for each array, we pass Array of Objects
+      if (decodedComponentType == Unknown) {
+        TypedClass(klass, parameters)
+      } else {
+        parameters match {
+          //it may happen that parameter will be decoded via other means, we have to to sanity check if they match
+          case Nil | `decodedComponentType` :: Nil =>
+            Typed.typedClass(classOf[Array[Object]], List(decodedComponentType))
+          case _: List[TypingResult] =>
+            throw new IllegalArgumentException(s"Array parameter passed twice, klass component type: ${klass.getComponentType}, type passed from parameters: ${parameters.head.display}")
+        }
+      }
+    }
 
     def genericTypeClass(klass: Class[_], params: List[TypingResult]): TypingResult = TypedClass(klass, params)